refactor amp/bento-selector
