+++ extensions/amp-selector/1.0/amp-selector.js
@@ -4,19 +4,32 @@ import {getWin} from '#core/window';
 
 import {isExperimentOn} from '#experiments';
 
+import {AmpPreactBaseElement} from '#preact/amp-base-element';
+
 import {Services} from '#service';
 
 import {createCustomEvent} from '#utils/event-helper';
 import {userAssert} from '#utils/log';
 
-import {BaseElement} from './base-element';
+import {Component, detached, elementInit, getOptions, props} from './element';
 
 import {CSS} from '../../../build/amp-selector-1.0.css';
 
 /** @const {string} */
 const TAG = 'amp-selector';
 
-class AmpSelector extends BaseElement {
+class AmpSelector extends AmpPreactBaseElement {
+  /** @override */
+  constructor(element) {
+    super(element);
+
+    /** @protected */
+    this.optionState = [];
+
+    /** @protected */
+    this.isExpectedMutation = false;
+  }
+
   /** @override */
   init() {
     // Set up API
@@ -46,7 +59,46 @@ class AmpSelector extends BaseElement {
       }
     });
 
-    return super.init();
+    return this.selectorInit();
+  }
+
+  /**
+   * @protected
+   * @return {JsonObject}
+   */
+  selectorInit() {
+    const props = elementInit(
+      this.element,
+      (mu) => {
+        if (this.isExpectedMutation) {
+          this.isExpectedMutation = false;
+          return;
+        }
+        const {children, options} = getOptions(this.element, mu);
+        this.optionState = options;
+        this.mutateProps({children, options});
+      },
+
+      // TODO(wg-bento): This hack is in place to prevent doubly rendering.
+      // See https://github.com/ampproject/amp-react-prototype/issues/40.
+      (event) => {
+        const {option, value} = event;
+        this.triggerEvent(
+          this.element,
+          'select',
+          dict({
+            'targetOption': option,
+            'selectedOptions': value,
+          })
+        );
+
+        this.isExpectedMutation = true;
+        this.mutateProps(dict({'value': value}));
+      }
+    );
+
+    this.optionState = props.options;
+    return props;
   }
 
   /** @override */
@@ -76,6 +128,15 @@ class AmpSelector extends BaseElement {
     return true;
   }
 }
+
+/** @override */
+AmpSelector['Component'] = Component;
+
+/** @override */
+AmpSelector['detached'] = detached;
+
+/** @override */
+AmpSelector['props'] = props;
 
 AMP.extension(TAG, '1.0', (AMP) => {
   AMP.registerElement(TAG, AmpSelector, CSS);
